chore: move useAuthenticated hook to remove circular dependency
