+++ site/src/hooks/index.ts
@@ -2,3 +2,4 @@ export * from "./useClickable";
 export * from "./useClickableTableRow";
 export * from "./useClipboard";
 export * from "./usePagination";
+export * from "./useAuthenticated";
