+++ site/src/modules/dashboard/DeploymentBanner/DeploymentBanner.tsx
@@ -1,6 +1,6 @@
 import { health } from "api/queries/debug";
 import { deploymentStats } from "api/queries/deployment";
-import { useAuthenticated } from "contexts/auth/RequireAuth";
+import { useAuthenticated } from "hooks";
 import type { FC } from "react";
 import { useQuery } from "react-query";
 import { DeploymentBannerView } from "./DeploymentBannerView";
