+++ site/src/pages/CreateWorkspacePage/CreateWorkspacePageExperimental.tsx
@@ -12,7 +12,7 @@ import type {
 	Workspace,
 } from "api/typesGenerated";
 import { Loader } from "components/Loader/Loader";
-import { useAuthenticated } from "contexts/auth/RequireAuth";
+import { useAuthenticated } from "hooks";
 import { useEffectEvent } from "hooks/hookPolyfills";
 import { generateWorkspaceName } from "modules/workspaces/generateWorkspaceName";
 import {
