+++ site/src/pages/UserSettingsPage/AccountPage/AccountPage.tsx
@@ -1,7 +1,7 @@
 import { groupsForUser } from "api/queries/groups";
 import { Stack } from "components/Stack/Stack";
 import { useAuthContext } from "contexts/auth/AuthProvider";
-import { useAuthenticated } from "contexts/auth/RequireAuth";
+import { useAuthenticated } from "hooks";
 import { useDashboard } from "modules/dashboard/useDashboard";
 import type { FC } from "react";
 import { useQuery } from "react-query";
