+++ site/src/pages/UserSettingsPage/Layout.tsx
@@ -1,7 +1,7 @@
 import { Loader } from "components/Loader/Loader";
 import { Margins } from "components/Margins/Margins";
 import { Stack } from "components/Stack/Stack";
-import { useAuthenticated } from "contexts/auth/RequireAuth";
+import { useAuthenticated } from "hooks";
 import { type FC, Suspense } from "react";
 import { Helmet } from "react-helmet-async";
 import { Outlet } from "react-router-dom";
